Add a part using z3c.recipe.compattest
This creates an isolated test runner for each named egg, including only it, its deps, and its test extra deps. Plus a script to run all such tests in sequence.

This is a good way to isolate issues that arise from running all the tests in the same process.

Some immediately apparent issues that show up are implicit dependencies that aren't specified. For example, nti.common is missing a dependency on itsdangerous. nti.dataserver seems to be missing a dependency on nti.contentlibrary.

This part isn't added to the parts list so you have to install it manually right now.